+++ cookbooks/hello-win-world/metadata.rb
@@ -1,0 +1,3 @@
+name 'hello-win-world'
+description 'Hello-World example for Windows, that created a file with some node infos'
+version '1.0.0'