+++ cookbooks/hello-win-world/recipes/default.rb
@@ -1,0 +1,5 @@
+template "c:/tmp/hello-world.txt" do
+  source "hello-file.erb"
+  mode "0644"
+  action :create
+end