+++ provision/assets/chef/client_win.rb
@@ -8,12 +8,12 @@ log_location            STDOUT
 exit_status             :enabled
 chef_server_url         "https://#{CHEF_SERVER_NAME}/organizations/#{ORGA_NAME}"
 validation_client_name  "#{ORGA_NAME}-validator"
-validation_key          "/etc/chef/#{ORGA_NAME}_org.key"
-ssl_ca_file             '/etc/chef/chef-server-tls.crt'
+validation_key          "#{DIR}/#{ORGA_NAME}_org.key"
+ssl_ca_file             "#{DIR}/chef-server-tls.crt"
 
-file_cache_path         CHEF_NODE_DIR_CACHE
-file_backup_path        CHEF_NODE_DIR_BACKUP
-cache_options( :path => CHEF_NODE_DIR_CHECKSUMS, :skip_expires => true )
+file_cache_path         "#{DIR}/cache"
+file_backup_path        "#{DIR}/backup"
+cache_options( :path => "#{DIR}/checksums", :skip_expires => true )
 
 data_collector.mode     :client
 
@@ -21,8 +21,8 @@ interval                10    # seconds, default: 1800
 
 #user                    CHEF_NODE_USER
 #group                   CHEF_NODE_GROUP
-cache_path              "/home/#{CHEF_NODE_USER}"
+cache_path               "c:/Users/chef"
 
 #use_policyfile          true
 #policy_name             'hello-world'   # role
-#policy_group            'playground'  # environment
+#policy_group            'winground'     # environment