+++ CHANGELOG.md
@@ -2,6 +2,12 @@ Changelog
 ==========
 
 
+### 1.1.0 - windows nodes
+
++   added support for windows nodes (Win Server 2012R2)
++   added win hello-world example
+
+
 ### 1.0.0 - initial commit
 
 +   simple hello-world example